ETS-Server/requirements.txt
paho-mqtt
pandas
# numpy
mysql-connector-python
PyYAML==5.4.1